requirements.txt
africastalking==1.2.4
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
aniso8601==9.0.1
APScheduler==3.10.0
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autocommand==2.2.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==23.1.0
blinker==1.6.2
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
cheroot==9.0.0
CherryPy==18.8.0
chromedriver-binary==114.0.5735.90.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.52.0
django-bootstrap-v5==1.0.11
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.3
django-filter==22.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-rest-authtoken==2.1.4
django-smart-selects==1.6.0
django-storages==1.13.2
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
dnspython==2.3.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Extractor==0.5
Flask==2.3.2
frozenlist==1.3.3
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflect==6.0.2
itsdangerous==2.1.2
itypes==1.2.0
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
Jinja2==3.1.2
jsonfield==3.1.0
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient @ file:///C:/Users/Satish/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl#sha256=1f4d3c3fa7d7f1683071ad12096e5494179e7fa837c5210c3032c81d72f33e3a
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.7
openpyxl==3.1.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
portend==3.1.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.3
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
schema==0.7.5
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tempora==5.2.1
tesseract==0.1.3
text-unidecode==1.3
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
twilio==7.16.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
whitenoise==6.3.0
wsproto==1.2.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.13.0
